# docs/Eventix-API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Eventix API",
    "description": "Eventix Event Booking System - Session-based auth, events, bookings, audit log. Import this file into Apifox or ApiDog for full API documentation.",
    "version": "1.1.0",
    "contact": {
      "name": "Eventix API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Eventix API / Auth"],
        "summary": "Refresh Token",
        "description": "Exchange refresh token for new access and refresh tokens. Session must be active and not expired.",
        "operationId": "refresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              },
              "example": {
                "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/AuthResponse" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "error": {
                    "code": "EVB401001",
                    "message": "Session expired or invalid",
                    "details": null
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{eventId}/bookings": {
      "post": {
        "tags": ["Eventix API / Bookings"],
        "summary": "Book Tickets",
        "description": "Book tickets for an event. Event must be published. Limits: max_tickets_per_booking = max per single request (prevents grabbing 50 at once); max_tickets_per_user = total tickets user can have for this event (user may need multiple requests to reach it). Default config: 10 per booking, 10 per user. Per-event overrides in event_booking_config.",
        "operationId": "bookSpot",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ticket_count": { "type": "integer", "minimum": 1, "maximum": 100, "default": 1, "description": "Number of tickets to book (1-100, subject to event config)" }
                },
                "example": { "ticket_count": 2 }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Booking" }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": 1,
                    "event_id": 1,
                    "user_id": 1,
                    "ticket_count": 2,
                    "status": "confirmed",
                    "created_at": "2025-03-06T10:00:00.000Z",
                    "updated_at": "2025-03-06T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (invalid event ID or ticket_count)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "error": { "code": "EVB400001", "message": "Validation failed", "details": { "ticket_count": "Number must be less than or equal to 100" } } }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "error": { "code": "EVB401001", "message": "Authentication required", "details": null } }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "error": { "code": "EVB404001", "message": "Event not found", "details": { "event_id": 999 } } }
              }
            }
          },
          "409": {
            "description": "Event sold out, not enough spots, or booking limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "soldOut": {
                    "summary": "Sold out",
                    "value": { "success": false, "error": { "code": "EVB409001", "message": "This event is sold out", "details": { "event_id": 1 } } }
                  },
                  "notEnoughSpots": {
                    "summary": "Not enough spots",
                    "value": { "success": false, "error": { "code": "EVB409001", "message": "Not enough spots available", "details": { "event_id": 1, "requested": 5, "remaining": 2 } } }
                  },
                  "limitExceeded": {
                    "summary": "Booking limit exceeded",
                    "value": { "success": false, "error": { "code": "EVB409003", "message": "You can book at most 10 tickets for this event (you have 8)", "details": { "max_tickets_per_user": 10, "current": 8 } } }
                  },
                  "perBookingLimit": {
                    "summary": "Per-booking limit exceeded",
                    "value": { "success": false, "error": { "code": "EVB409003", "message": "Cannot book more than 10 tickets per booking", "details": { "max_tickets_per_booking": 10 } } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings": {
      "get": {
        "tags": ["Eventix API / Bookings"],
        "summary": "List Bookings",
        "description": "List bookings. User: own bookings only. Admin: all bookings.",
        "operationId": "listBookings",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 }, "description": "Items per page (max 100)" }
        ],
        "responses": {
          "200": {
            "description": "List of bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Booking" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "description": "Authentication required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/bookings/{bookingId}": {
      "get": {
        "tags": ["Eventix API / Bookings"],
        "summary": "Get Booking by ID",
        "description": "Get a booking by ID. User: own only. Admin: any.",
        "operationId": "getBookingById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "bookingId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Booking details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Booking" }
                  }
                }
              }
            }
          },
          "401": { "description": "Authentication required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "error": { "code": "EVB403001", "message": "You do not have permission to perform this action", "details": null } }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "error": { "code": "EVB404001", "message": "Booking not found", "details": { "booking_id": 999 } } }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Eventix API / Bookings"],
        "summary": "Cancel Booking",
        "description": "Cancel a booking. Accepts empty body or { status: 'cancelled' }.",
        "operationId": "cancelBooking",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "bookingId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["cancelled"], "example": "cancelled" }
                }
              },
              "example": { "status": "cancelled" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Booking" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed (invalid body)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "400": { "description": "Validation failed (invalid body)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "400": { "description": "Validation failed (invalid body)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "400": { "description": "Validation failed (invalid body)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Authentication required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Booking not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "400": { "description": "Validation failed (invalid body)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "400": { "description": "Validation failed (invalid body)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": {
            "description": "Booking already cancelled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "success": false, "error": { "code": "EVB422001", "message": "Booking is already cancelled", "details": { "booking_id": 1 } } }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login or register"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "EVB400001" },
              "message": { "type": "string", "example": "Validation failed" },
              "details": { "type": "object", "nullable": true }
            }
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "email": { "type": "string" },
              "name": { "type": "string", "nullable": true }
            }
          },
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "expires_in": { "type": "integer", "description": "Access token expiry in seconds" }
        }
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "event_id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "ticket_count": { "type": "integer", "description": "Number of tickets in this booking" },
          "status": { "type": "string", "enum": ["confirmed", "cancelled"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_prev": { "type": "boolean" }
        }
      }
    }
  }
}